+++ spec/02-integration/03-db/03-plugins_spec.lua
@@ -206,56 +206,6 @@ for _, strategy in helpers.each_strategy() do
         assert.falsy(ok)
         assert.match("missing plugin is enabled but not installed", err, 1, true)
       end)
-
-      it("reports failure with bad plugins #4392", function()
-        local ok, err = db.plugins:load_plugin_schemas({
-          ["legacy-plugin-bad"] = true,
-        })
-        assert.falsy(ok)
-        assert.match("failed converting legacy schema for legacy-plugin-bad", err, 1, true)
-      end)
-
-      it("succeeds with good plugins", function()
-        local ok, err = db.plugins:load_plugin_schemas({
-          ["legacy-plugin-good"] = true,
-        })
-        assert.truthy(ok)
-        assert.is_nil(err)
-
-        local foo = {
-          required = false,
-          type = "map",
-          keys = { type = "string" },
-          values = { type = "string" },
-          default = {
-            foo = "boo",
-            bar = "bla",
-          }
-        }
-        local config = {
-          type = "record",
-          required = true,
-          fields = {
-            { foo = foo },
-            foo = foo,
-          }
-        }
-        local consumer = {
-          type = "foreign",
-          reference = "consumers",
-          eq = ngx.null,
-          schema = db.consumers.schema,
-        }
-        assert.same({
-          name = "legacy-plugin-good",
-          fields = {
-            { config = config },
-            { consumer = consumer },
-            config = config,
-            consumer = consumer,
-          }
-        }, db.plugins.schema.subschemas["legacy-plugin-good"])
-      end)
     end)
   end) -- kong.db [strategy]
 end