+++ spec/fixtures/custom_plugins/kong/plugins/dummy/schema.lua
@@ -1,12 +1,15 @@
 return {
   name = "dummy",
   fields = {
-    { config = {
+    {
+      config = {
         type = "record",
         fields = {
-          { resp_header_value = { type = "string", default = "1" }, },
-          { append_body = { type = "string" }, },
-          { resp_code = { type = "number" }, },
-    }, }, },
+          { resp_header_value = { type = "string", default = "1" } },
+          { append_body = { type = "string" } },
+          { resp_code = { type = "number" } },
+        },
+      },
+    },
   },
 }