+++ spec/fixtures/custom_plugins/kong/plugins/legacy-plugin-good/handler.lua
@@ -1,7 +1,0 @@
-local LegacyPluginGoodHandler =  {
-  VERSION = "0.1-t",
-  PRIORITY = 1000,
-}
-
-
-return LegacyPluginGoodHandler