+++ CHANGELOG.md
@@ -89,6 +89,9 @@
   or the typed `shorthand_fields`. If your custom schemas still use `shorthands`, you
   need to update them to use `shorhand_fields`.
   [#8815](https://github.com/Kong/kong/pull/8815)
+- The support for deprecated legacy plugin schemas was removed. If your custom plugins
+  still use the old (`0.x era`) schemas, you are now forced to upgrade them.
+  [#8815](https://github.com/Kong/kong/pull/8815)
 
 #### Admin API